changed relevant datatypes to size_t and ssize_t to always fit the architecture

+++ userspace/libc/include/wait.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "types.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -9,7 +11,7 @@ extern "C" {
 //pid typedefs
 #ifndef PID_T_DEFINED
 #define PID_T_DEFINED
-typedef long int pid_t;
+typedef ssize_t pid_t;
 #endif // PID_T_DEFINED
 
 extern pid_t waitpid(pid_t pid, int *status, int options);
